+++ NEWS.md
@@ -1,7 +1,7 @@
 # lofifonts 0.1.0.9001  2024-10-29
 
 * [9000] Optimised the storage and rendering speed
-* [9001] Convert bittermelon fonts to 'lofifont' spec
+* [9001] Convert bittermelon fonts to 'lofi' font
 
 # lofifonts 0.1.0